Waiter UI Database implementation

+++ server/src/endpoints/order/OrderRequestParameters.java
@@ -2,13 +2,13 @@ package endpoints.order;
 
 public class OrderRequestParameters {
 
-  private Long tableNumber;
+  private int tableNumber;
 
-  public OrderRequestParameters(Long tableNumber) {
+  public OrderRequestParameters(int tableNumber) {
     this.tableNumber = tableNumber;
   }
 
-  public Long getTableNumber() {
+  public int getTableNumber() {
     return tableNumber;
   }
 }
